% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_site.R
\name{commit_site}
\alias{commit_site}
\title{Commit the website files}
\usage{
commit_site(dry_run = FALSE, path = ".")
}
\arguments{
\item{dry_run}{Identifies R Markdown files that have been updated, but does
not render them.}

\item{path}{By default the function assumes the current working directory is
within the project. If this is not true, you'll need to provide the path to
the project directory.}
}
\value{
If \code{dry_run = TRUE}, returns the character vector of R Markdown
  files that would be rendered. Otherwise invisibly returns this vector.
}
\description{
\code{commit_site} ensures that the website files are created by the code
that has been committed to the Git repository.
}
\examples{
\dontrun{
commit_site()
}
}

