% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_options.R
\name{create_figure_path}
\alias{create_figure_path}
\title{create the path to the figure folder}
\usage{
create_figure_path(input)
}
\arguments{
\item{input}{the path to an R Markdown file}
}
\description{
create the path to the figure folder
}
\keyword{internal}
