% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{create_gitignore}
\alias{create_gitignore}
\title{Create a default .gitignore file}
\usage{
create_gitignore(path, overwrite = FALSE)
}
\description{
The .gitignore in inst/infrastrucure does not survive builing the R package.
The .nojekyll does, so it must be specific to this filename and not a
property of hidden files. Hadley does not include .gitignore in
.Rbuildignore, which further supports that it is ignored by default.
}

