% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_commit.R
\name{decide_to_render}
\alias{decide_to_render}
\title{Decide with files to render and commit}
\usage{
decide_to_render(repo, log, rmd)
}
\description{
Recursively search the commit log until the R Markdown file or its
corresponding HTML file is found. If the Rmd is found first, the HTML file
needs to be re-rendered, added, and committed (return TRUE). If the HTML file
is found first, then it is up-to-date (return FALSE).
}
\seealso{
\code{\link{obtain_files_in_commit}},
  \code{\link{obtain_files_in_commit_root}}, \code{\link{wflow_commit}}
}

