% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_options.R
\name{is_fig_path_ext}
\alias{is_fig_path_ext}
\title{check the fig_path_ext option}
\usage{
is_fig_path_ext(input)
}
\arguments{
\item{input}{the path to an R Markdown file}
}
\description{
check the fig_path_ext option
}
\keyword{internal}
