% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_site.R
\name{obtain_files_in_commit_root}
\alias{obtain_files_in_commit_root}
\title{Obtain the files updated in the root commit}
\usage{
obtain_files_in_commit_root(repo, commit)
}
\description{
There is currently no obvious way to obtain the files included in the initial
root commit using the git2r API. Thus these are obtained using the
command-line. See
\url{http://stackoverflow.com/questions/41433034/how-to-obtain-files-included-in-initial-commit-using-git2r-libgit2}.
}
\details{
This only works for the root commit, i.e. it must have no parents.
}
\seealso{
\code{\link{obtain_files_in_commit}}, \code{\link{decide_to_render}}
}

