% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_project.R
\name{start_project}
\alias{start_project}
\title{Start a new workflowr project.}
\usage{
start_project(name, directory, git_init = TRUE, rstudio = TRUE)
}
\arguments{
\item{name}{Project name, e.g. "My Project"}

\item{directory}{The directory for the project, e.g. "new-project". Will be
created if necessary.}

\item{git_init}{Control whether the directory is initiatlized as a Git
repository (default: TRUE).}

\item{rstudio}{Control whether an RStudio Project file is created (default:
TRUE).}
}
\description{
\code{start_project} creates a new project directory for the site.
}
\details{
This is the initial function that organizes the infrastructure to create a
research website for your project.
}
\examples{
\dontrun{
start_project("My Project", "path/to/new-project")
}
}

