% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_archive.R
\name{wflow_archive}
\alias{wflow_archive}
\title{Archive objects}
\usage{
wflow_archive(..., archive_dir = "../archive", subdir = NULL,
  overwrite = FALSE, path = ".")
}
\arguments{
\item{...}{Objects to be archived.}

\item{archive_dir}{character (default: "../archive"). The path to the main
archive directory.}

\item{subdir}{character (default: NULL). The name of the subdirectory to save
the objects. If NULL, the name of the R Markdown file is used.}

\item{overwrite}{logical (default: FALSE). If the object already has an
archived file for the current version of the workflowr project, should it
be overwritten?}

\item{path}{character (default: "."). By default the function assumes the
current working directory is within the project. If this is not true,
you'll need to provide the path to the project directory.}
}
\description{
\code{wflow_archive} archives objects from a specific version (i.e. Git
commit) of the workflowr project.
}
\details{
\code{wflow_archive} serializes each object to a separate file using
\code{\link{saveRDS}}. To maintain organization and reduce the possibility of
conflict, the default is to save the objects within a subdirectory of the
archive directory which is named after the R Markdown file being knit (i.e.
the path is \code{archive_dir/subdir}).

Note that this function uses non-standard evaluation to save the objects. Do
not use \code{wflow_archive} within a function (please open an Issue if it
would be useful to you to have a standard evaluation option for use within
functions).
}
\examples{
\dontrun{
x <- 1
y <- 2
z <- 3
wflow_archive(x, y, z)
}
}
\seealso{
\code{\link{wflow_restore}}, \code{\link{saveRDS}}
}

