% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_convert.R
\name{wflow_convert}
\alias{wflow_convert}
\title{Convert R Markdown files to workflowr template.}
\usage{
wflow_convert(files, standalone = FALSE, dry_run = FALSE, verbose = TRUE)
}
\arguments{
\item{files}{character. The R Markdown file(s) to be converted. Must have
file extension Rmd or rmd.}

\item{standalone}{logical (default: FALSE). Embed the configuration chunks
directly in the document to create a standalone file. Should only need to
be used when creating files outside the context of a workflowr project.}

\item{dry_run}{logical (default: FALSE). Preview the changes to the files.
Show the output of \code{diff} between the original and the proposed
replacement.}

\item{verbose}{logical (default: TRUE). Report progress.}
}
\value{
Invisibly returns the files which were successfully converted.
}
\description{
\code{wflow_convert} converts existing R Markdown files to use the workflowr
template. This is especially useful when migrating an existing project with
many R Markdown files to use workflowr.
}
\section{Warning}{
 \code{wflow_convert} overwrites the original files.
}

\examples{
\dontrun{

# Convert an R Markdown file to use workflowr template
wflow_convert("existing.Rmd")
# Convert an R Markdown file to use workflowr template,
# but embed the shared chunks to create a standalone file
wflow_convert("existing.Rmd", standalone = TRUE)
# Preview the changes
wflow_convert("existing.Rmd", dry_run = TRUE)
# Convert multiple an R Markdown files
wflow_convert(files = list.files(path = "analysis",
                                 pattern = "Rmd$",
                                 full.names = TRUE))
}

}
\seealso{
\code{\link{wflow_open}}
}
