% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_html.R
\name{wflow_html}
\alias{wflow_html}
\title{Convert to a workflowr HTML document}
\usage{
wflow_html(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[rmarkdown]{html_document}}.}
}
\value{
An \code{\link[rmarkdown]{output_format}} object to pass to
\code{\link[rmarkdown]{render}}.
}
\description{
Workflowr custom format for converting from R Markdown to an HTML
document. See \code{\link[rmarkdown]{render}}
\code{\link[rmarkdown]{render_site}} for an explanation of how this
conversion format can be specified inside the \code{_site.yml}
configuration file or an R Markdown file. This is intended to be
used to generate webpages for a workflowr website, but it can also
be used outside a workflowr project to generate webpages from R
Markdown documents.
}
