% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_html.R
\name{wflow_post_knit}
\alias{wflow_post_knit}
\title{post_knit function for workflowr}
\usage{
wflow_post_knit(metadata, input_file, runtime, encoding, ...)
}
\arguments{
\item{metadata}{The metadata specified in the YAML header of the R Markdown
file}

\item{input_file}{Name of R Markdown file}

\item{runtime}{The runtime target for rendering. The \code{static} option
produces output intended for static files; \code{shiny} produces output
suitable for use in a Shiny document (see \code{\link[rmarkdown]{run}}). The default,
\code{auto}, allows the \code{runtime} target specified in the YAML metadata
to take precedence, and renders for a \code{static} runtime target otherwise.}

\item{encoding}{Ignored. The encoding is always assumed to be UTF-8.}

\item{...}{arguments passed to the \code{post_knit} function of
\code{rmarkdown::\link[rmarkdown]{html_document}}}
}
\description{
This is the \code{post_knit} function that \code{\link{wflow_html}} passes to
the function \code{\link[rmarkdown]{output_format}} from the package
\link{rmarkdown}. For advanced usage only.
}
\details{
If you'd like to combine workflowr with another R Markdown output format, you
may need to use \code{wflow_post_knit}. This function fixes the path to the R
Markdown file (which is manipulated by \code{\link{wflow_pre_knit}}).
}
\seealso{
\code{\link{wflow_html}}, \code{\link{wflow_pre_knit}},
         \code{\link{wflow_pre_processor}}
}
\keyword{internal}
