% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_html.R
\name{wflow_pre_knit}
\alias{wflow_pre_knit}
\title{pre_knit function for workflowr}
\usage{
wflow_pre_knit(input, ...)
}
\arguments{
\item{input}{Name of R Markdown file}

\item{...}{currently ignored}
}
\description{
This is the \code{pre_knit} function that \code{\link{wflow_html}} passes to
the function \code{\link[rmarkdown]{output_format}} from the package
\link{rmarkdown}. For advanced usage only.
}
\details{
If you'd like to insert the workflowr reproducibility report into other R
Markdown output formats such as \code{blogdown::html_page}, you can use
\code{wflow_pre_knit}.
}
\seealso{
\code{\link{wflow_html}}, \code{\link{wflow_post_knit}},
         \code{\link{wflow_pre_processor}}
}
\keyword{internal}
