% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_start.R
\name{wflow_start}
\alias{wflow_start}
\title{Start a new workflowr project.}
\usage{
wflow_start(name, directory, git = TRUE, existing = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{name}{character. Project name, e.g. "My Project"}

\item{directory}{character. The directory for the project, e.g.
"~/new-project". By default the directory will be created (\code{existing =
FALSE}.}

\item{git}{logical (default: TRUE). Should Git be used for version control?
If \code{directory} is a new Git repository, \code{wflow_start} will
initialize the repository and make an initial commit. If \code{directory}
is already a Git repository, \code{wflow_start} will make an additional
commit. In both cases, only files needed for the workflowr project will be
included in the commit.}

\item{existing}{logical (default: FALSE). Indicate if the specified
\code{directory} already exists. The default prevents injecting the
workflowr files into an unwanted location. Only set to TRUE if you wish to
add the workflowr files to an existing project.}

\item{overwrite}{logical (default: FALSE). Control whether to overwrite
existing files. Only relevant if \code{existing = TRUE}. Passed to
\code{file.copy}.}
}
\description{
\code{wflow_start} adds the necesary files for a workflowr project. The
default is to create a new directory, but it is also possible to add the
files to an already existing project.
}
\details{
This is the initial function that organizes the infrastructure to create a
research website for your project. Note that while you do not need to use
RStudio with workflowr, do not delete the Rproj file because it is required
by other functions.
}
\examples{
\dontrun{
wflow_start("My Project", "path/to/new-project")
}
}
