% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_update.R
\name{wflow_update}
\alias{wflow_update}
\title{Update a workflowr project.}
\usage{
wflow_update(dry_run = TRUE, commit = TRUE, log_file = NULL,
  log_open = interactive(), path = ".")
}
\arguments{
\item{dry_run}{logical (default: TRUE). Preview the proposed updates.}

\item{commit}{logical (default: TRUE). Commit the updated files. Only
executed if \code{dry_run = FALSE}.}

\item{log_file}{character (default: NULL). A file to save the log messages.
If NULL, a temporary file is created.}

\item{log_open}{logical (default: \code{interactive()}). Should the log file
be opened in RStudio? This argument is ignored if the function is not run
from within RStudio.}

\item{path}{character (default: ".") By default the function assumes the
current working directory is within the project. If this is not true,
you'll need to provide the path to the project directory.}
}
\value{
A character vector of the updated files.
}
\description{
Newer versions of workflowr sometimes make changes that need to be
coordinated across multiple files. After upgrading workflowr, run
\code{wflow_update} to make all the necessary changes.
}
\details{
By default, \code{wflow_update} is run in \code{dry_run} mode so that no
unwanted changes are made. The log files contains the changes to each file,
represented with the syntax from the
\href{https://en.wikipedia.org/wiki/Diff_utility#Usage}{Unix diff utility}.
After reviewing the log file for the proposed changes, re-run the function
with \code{dry_run = FALSE} to implement them.

Currently \code{wflow_update} checks for the following items:

\itemize{

\item Updates the shared chunks in \code{analysis/chunks.R}.

\item Updates each R Markdown file in \code{analysis/} to use the shared
chunks. This is implemented with \code{\link{wflow_convert}}.

\item Removes the "BuildType: Website" from the Rproj file. This re-builds
every R Markdown file everytime, so it is safer to always use
\code{\link{wflow_build}}.

}
}
\examples{
\dontrun{

# Preview the potential changes
wflow_update()
# Incorporate the changes
wflow_update(dry_run = FALSE)
}

}
\seealso{
\code{\link{wflow_convert}}
}
