% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_build.R
\name{wflow_build}
\alias{wflow_build}
\title{Build the site}
\usage{
wflow_build(files = NULL, make = is.null(files), update = FALSE,
  republish = FALSE, seed = 12345, log_dir = NULL, local = FALSE,
  dry_run = FALSE, project = ".")
}
\arguments{
\item{files}{character (default: NULL). Files to build. Supports file
extensions Rmd, rmd, and md. Only files in the analysis directory are
allowed (and therefore any path to a file is ignored).}

\item{make}{logical (default: \code{is.null(files)}). Use Make-like behavior,
i.e. build the files that have been updated more recently than their
corresponding HTML files. This is the default action if no files are
specified.}

\item{update}{logical (default: FALSE). Build the files that have been
committed more recently than their corresponding HTML files (and do not
have any unstaged or staged changes). This ensures that the commit version
ID inserted into the HTML corresponds to the exact version of the source
file that was used to produce it.}

\item{republish}{logical (default: FALSE). Build all published R Markdown
files. Useful for site-wide changes like updating the theme, navigation
bar, or any other setting in \code{_site.yml}.}

\item{seed}{numeric (default: 12345). The seed to set before building each
file. Passed to \code{\link{set.seed}}.}

\item{log_dir}{character (default: NULL). The directory to save log files
from building files. It will be created if necessary and ignored if
\code{local = TRUE}. The default is to create a directory in \code{/tmp}.}

\item{local}{logical (default: FALSE). Build files locally in the R console.
This should only be used for debugging purposes. The default is to build
each file in its own separate fresh R process to ensure each file is
reproducible in isolation.}

\item{dry_run}{logical (default: FALSE). Preview the files to be built, but
do not actually build them.}

\item{project}{character (default: ".") By default the function assumes the
current working directory is within the project. If this is not true,
you'll need to provide the path to the project directory.}
}
\value{
An object of class \code{wflow_build}, which is a list with the
  following elements:

 \itemize{

   \item \bold{files}: The input argument \code{files}

   \item \bold{make}: The input argument \code{make}

   \item \bold{update}: The input argument \code{update}

   \item \bold{republish}: The input argument \code{republish}

   \item \bold{seed}: The input argument \code{seed}

   \item \bold{log_dir}: The directory where the log files were saved

   \item \bold{local}: The input argument \code{files}

   \item \bold{dry_run}: The input argument \code{files}

   \item \bold{built}: The relative paths to the built R Markdown files

   \item \bold{html}: The relative paths to the corresponding HTML files

 }
}
\description{
\code{wflow_build} builds the website from the files in the analysis
directory. This is intended to be used when developing your code to preview
the changes. When you are ready to commit the files, use
\code{\link{wflow_publish}}.
}
\details{
\code{wflow_build} has multiple, non-mutually exclusive options for deciding
which files to build:

\itemize{

\item Files specified via the argument \code{files} are always built.

\item If \code{make = TRUE}, all files which have been modified more recently
than their corresponding HTML files will be built.

\item If \code{update = TRUE}, all files which have been committed more
recently than their corresponding HTML files will be built. However, files
which currently have staged or unstaged changes will be ignored.

\item If \code{republish = TRUE}, all published files will be rebuilt.

}

Under the hood, \code{wflow_build} is a wrapper for
\code{\link[rmarkdown]{render_site}} from the package \link{rmarkdown}.
}
\examples{
\dontrun{

# Build all files
wflow_build() # equivalent to wflow_build(make = TRUE)
# Build a single file
wflow_build("file.Rmd")
# Build multiple files
wflow_build(c("file1.Rmd", "file2.Rmd"))
# Build every file
wflow_build(republish = TRUE)
}

}
\seealso{
\code{\link{wflow_publish}}
}
