% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_build.R
\name{wflow_build}
\alias{wflow_build}
\title{Build the website}
\usage{
wflow_build(all = FALSE, files = NULL, dry_run = FALSE, path = ".", ...)
}
\arguments{
\item{all}{logical indicating if every R Markdown file should be rendered
when building the site (default: FALSE).}

\item{files}{R Markdown files to be rendered. The files can be specified
using the path or just the basename (default: NULL).}

\item{dry_run}{Identifies R Markdown files that have been updated, but does
not render them.}

\item{path}{By default the function assumes the current working directory is
within the project. If this is not true, you'll need to provide the path to
the project directory.}

\item{...}{Additional arguments that can be passed to
\code{rmarkdown::render_site}. Should only be needed for testing potential
changes. Any permanent settings should be specified in
\code{analysis/_site.yml}.}
}
\value{
If \code{dry_run = TRUE}, returns the character vector of R Markdown
  files that would be rendered. Otherwise invisibly returns this vector.
}
\description{
\code{wflow_build} builds the website by rendering the R Markdown files in
the analysis directory and saving them to \code{docs/}. By default it only
renders the R Markdown files that have been modified more recently than their
corresponding HTML file (similar to a Makefile). To render every single page
(e.g. to change the theme across the entire site), set \code{all = TRUE}. To
render specific R Markdown files, pass them as a vector to the argument
\code{files}.
}
\details{
Under the hood, this runs \code{rmarkdown::render_site} on each updated file
individually. This provides all the website styling specified in
\code{_site.yml} (which you would lose if you ran \code{rmarkdown::render})
without re-building the entire site (which would happen if you ran
\code{rmarkdown::render_site with no arguments}).

To include R Markdown files in your workflowr project that are not included
as part of the website, you have multiple options: 1) Prepend an underscore
to the filename, 2) move them to a subdirectory within the analysis
directory, or 3) move them to another directory at the root of your project.
}
\examples{
\dontrun{
# View the files that would be rendered
wflow_build(dry_run = TRUE)
# Render all modified files
wflow_build()
# Render all files
wflow_build(all = TRUE)
# Render specific files
wflow_build(files = c("one.Rmd", "two.Rmd"))
}
}

