% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_html.R
\name{wflow_html}
\alias{wflow_html}
\title{Reproducible HTML document}
\usage{
wflow_html(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[rmarkdown]{html_document}}}
}
\value{
\code{\link[rmarkdown]{output_format}}
}
\description{
The output format \code{wflow_html} automatically 1) sets a seed with
\code{\link{set.seed}}, 2) inserts version of Git repo, and 3) inserts
\code{\link{sessionInfo}}.
}
