% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_archive.R
\name{wflow_restore}
\alias{wflow_restore}
\title{Restore objects}
\usage{
wflow_restore(..., subdir, archive_dir = "../archive", path = ".")
}
\arguments{
\item{...}{Objects to be restored.}

\item{subdir}{character. The name of the subdirectory to load the objects.}

\item{archive_dir}{character (default: "../archive"). The path to the main
archive directory.}

\item{path}{character (default: "."). By default the function assumes the
current working directory is within the project. If this is not true,
you'll need to provide the path to the project directory.}
}
\value{
Returns a nested list. The top-level list has one element for each
  object that was restored. Each of these elements in turn are lists with one
  element per code version (i.e. Git commit ID).
}
\description{
\code{wflow_restore} restores all previous versions of objects which had been
archived with \code{wflow_archive}.
}
\details{
\code{wflow_restore} loads each serialized object to a separate file using
\code{\link{readRDS}}. It searches for the archived objects in
\code{archive_dir/subdir}).

Note that this function uses non-standard evaluation to save the objects. Do
not use \code{wflow_restore} within a function (please open an Issue if it
would be useful to you to have a standard evaluation option for use within
functions).
}
\examples{
\dontrun{
restored <- wflow_restore(x, y, z, subdir = "filename.Rmd")
}
}
\seealso{
\code{\link{wflow_archive}}, \code{\link{readRDS}}
}

