% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_start.R
\name{wflow_start}
\alias{wflow_start}
\title{Start a new workflowr project.}
\usage{
wflow_start(name, directory, git_init = TRUE, overwrite = FALSE)
}
\arguments{
\item{name}{Project name, e.g. "My Project"}

\item{directory}{The directory for the project, e.g. "new-project". Will be
created if necessary.}

\item{git_init}{Control whether the directory is initiatlized as a Git
repository (default: TRUE).}

\item{overwrite}{overwrite    Control whether to overwrite existing files
(default: FALSE). Passed to \code{file.copy}.}
}
\description{
\code{wflow_start} creates a new project directory for the site.
}
\details{
This is the initial function that organizes the infrastructure to create a
research website for your project. Note that while you do not need to use
RStudio with workflowr, do not delete the Rproj file because it is required
by other functions.
}
\examples{
\dontrun{
wflow_start("My Project", "path/to/new-project")
}
}

