% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_start.R
\name{wflow_start}
\alias{wflow_start}
\title{Start a new workflowr project.}
\usage{
wflow_start(name, directory, git = TRUE, overwrite = FALSE)
}
\arguments{
\item{name}{character. Project name, e.g. "My Project"}

\item{directory}{character. The directory for the project, e.g.
"~/new-project". Will be created if necessary.}

\item{git}{logical (default: TRUE). Should Git be used for version control?
If \code{directory} is a new Git repository, \code{wflow_start} will
initialize the repository and make an initial commit. If \code{directory}
is already a Git repository, \code{wflow_start} will make an additional
commit. In both cases, only files needed for the workflowr project will be
included in the commit.}

\item{overwrite}{logical (default: FALSE). Control whether to overwrite
existing files . Passed to \code{file.copy}.}
}
\description{
\code{wflow_start} creates a new project directory for the site.
}
\details{
This is the initial function that organizes the infrastructure to create a
research website for your project. Note that while you do not need to use
RStudio with workflowr, do not delete the Rproj file because it is required
by other functions.
}
\examples{
\dontrun{
wflow_start("My Project", "path/to/new-project")
}
}

