% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_status.R
\name{wflow_status}
\alias{wflow_status}
\title{Report status of workflowr project.}
\usage{
wflow_status(files = NULL, project = ".")
}
\arguments{
\item{files}{character (default: NULL) The analysis file(s) to
report the status. By default checks the status of all analysis
files.}

\item{project}{character (default: ".") By default the function
assumes the current working directory is within the project. If
this is not true, you'll need to provide the path to the project
directory.}
}
\value{
Returns an object of class \code{wflow_status}, which is a
  list with the following elements:

  \itemize{

  \item \bold{wd}: The current working directory in the R console
  (i.e. \code{getwd{}}).

  \item \bold{root}: The root directory of the workflowr project
  (i.e. contains the RStudio .Rproj file).

  \item \bold{analysis}: The directory that contains
  \code{_site.yml} and the R Markdown files.

  \item \bold{docs}: The directory that contains the HTML files and
  figures.

  \item \bold{files}: The files whose status was checked.

  \item \bold{git}: The \code{.git} directory that contains the
  history of the Git repository.

  \item \bold{git_status}: The output from
  \code{git2r::\link[git2r]{status}}.

  \item \bold{status}: A data frame with detailed information on
  the status of each file (see below).

  }

  The data frame \code{status} contains the following columns (all
  logical vectors):

  \itemize{

  \item \bold{outdated}: When an R Markdown file has been committed
  to the repository without updating the previously published HTML
  file.

  \item \bold{staged}: When an R Markdown file has changes that
  have been added to the index (e.g. with \code{git add}).

  \item \bold{unstaged}: When a tracked R Markdown file has changes
  in the working directory.

  \item \bold{untracked}: When an R Markdown file has not been
  added or committed to the Git repository.

  \item \bold{ignored}: When an R Markdown file has been ignored by
  Git according to the patterns in the file \code{.gitignore}.

  }
}
\description{
\code{wflow_status} reports the analysis files that require user
action.
}
\details{
\code{wflow_status} reports analysis files with one of the
following statuses:

\itemize{

\item \bold{Mod}: Modified file. Any published file that has been
modified since the last time the HTML was published.

\item \bold{Unp}: Unpublished file. Any tracked file whose
corresponding HTML is not tracked. May or may not have staged or
unstaged changes.

\item \bold{New}: New file. Any untracked file that is not
specifically ignored.

}

\code{wflow_status} only works for workflowr projects that use Git.
}
\examples{
\dontrun{

wflow_status()
# Get status of specific file(s)
wflow_status("analysis/file.Rmd")
# Save the results
s <- wflow_status()
}
}
