% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_use_gitlab.R
\name{wflow_use_gitlab}
\alias{wflow_use_gitlab}
\title{wflow_use_gitlab}
\usage{
wflow_use_gitlab(username = NULL, repository = NULL,
  navbar_link = TRUE, protocol = "https", domain = "gitlab.com",
  project = ".")
}
\arguments{
\item{project}{character (default: ".") By default the function assumes the
current working directory is within the project. If this is not true,
you'll need to provide the path to the project directory.}
}
\description{
Automate setup for deploying website with GitLab:
}
\details{
1. Rename docs/ to public/
2. Edit output_dir in _site.yml
3. Add link in navigation bar
4. Create .gitlab-ci.yml

https://docs.gitlab.com/ee/ci/yaml/README.html#pages
}
