% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{workflowr}
\alias{workflowr}
\alias{workflowr-package}
\title{workflowr: A workflow template for creating a research website}
\description{
The workflowr package helps you create a research website using R Markdown
and Git.
}
\section{Vignettes}{


  Run \code{browseVignettes("workflowr")} to read the package vignettes
  locally. Alternatively you can read the documentation online at
  \url{https://jdblischak.github.io/workflowr}.
}

\section{Main workflowr functions}{


  \itemize{

  \item \code{\link{wflow_start}} - Start workflowr project

  \item \code{\link{wflow_build}} - Build the site to view locally

  \item \code{\link{wflow_publish}} - Publish analyses to include in the
  website

  \item \code{\link{wflow_status}} - Report status of analysis files

  }
}

